+++ cmd/e2e_test.go
@@ -4554,6 +4554,10 @@ func (m baseMock) GetCloudSavings() (*client.CloudSavings, error) {
 	return nil, errors.New("not implemented")
 }
 
+func (m baseMock) GetCloudLedger() (*client.CloudLedger, error) {
+	return nil, errors.New("not implemented")
+}
+
 func (m baseMock) GetClusterCost(clusterID string) (*client.ClusterCost, error) {
 	return nil, errors.New("not implemented")
 }